Data from the anxiety and confinement study.

This dataset contains the variables used in the anxiety and confinement study carried out by Alvarado-Aravena et al. 2022.

Usage

anxiety

Format

A data frame with 617 rows y 7 variables:

id: Factor. An identification code for each subject.

sex: Factor w/ 2 levels "Female", "Male". Sex of participants.

zone: Factor w/ 2 levels "CZ", "UZ". Zone in which the subject were by the time he was answering the questionnaire, either CZ (Confinement Zone) or UZ (Unconfinement Zone).

beck_global: Integer. Global score of Beck Anxiety Inventory.

pits_global: Integer. Global score of Pittsburgh Sleep Quality Index.

age: Integer. Age of the subjects in years.

cat_age: Factor w/ 4 levels "18-25", "26-40", "41-50", ">50". Age of the subjects in years.

Source

Alvarado-Aravena, C., Arriaza, K., Castillo-Aguilar, M., Flores, K., Dagnino-Subiabre, A., Estrada-Goic, C., & Núñez-Espinosa, C. (2022). Effect of Confinement on Anxiety Symptoms and Sleep Quality during the COVID-19 Pandemic. Behavioral Sciences, 12(10), 398.

Examples

# Mean age grouped by sex and zone using `data.table` syntax
anxiety[,  # No filtering (i)
        list(mean_age = mean(age)), # Action to do (j)
        list(sex, zone)] # Grouping vars (by)
#>       sex zone mean_age
#> 1: Female   CZ 30.46222
#> 2:   Male   CZ 33.11290
#> 3: Female   PZ 40.24034
#> 4:   Male   PZ 39.69072

# Mean PSQI score grouped by sex and zone, for those with
# an age greater than 18 AND a Beck score greater than 10.
anxiety[age > 18 & beck_global > 10,
        list(mean_psqi = mean(pits_global)),
        list(sex, zone)]
#>       sex zone mean_psqi
#> 1: Female   CZ  10.66346
#> 2:   Male   CZ  12.00000
#> 3:   Male   PZ  10.00000
#> 4: Female   PZ  12.00000